Added copy/paste buttons to profile and layer list. Improved layer/profile copying so profiles can be copied across applications. Layers specific to a application (e.g. MinecraftRainLayer) are not copied. Added button to import json file from disk, so users do not have to manually do it.
